JunitMockitoExample/src/main/java/com/example/mockito/service/EmployeeService.java
package com.example.mockito.service;

import com.example.mockito.model.Employee;
import com.example.mockito.repository.EmployeeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class EmployeeService {
    @Autowired
    private EmployeeRepository employeeRepository;

    public Employee saveEmployee(Employee employee) {
        return employeeRepository.save(employee);
    }

    public List<Employee> getAllEmployees() {
        return employeeRepository.findAll();
    }

    public Employee getEmployeeById(String id) {
        return employeeRepository.findById(id).orElse(null);
    }

    public void deleteEmployeeById(String id) {
        employeeRepository.deleteById(id);
    }

    public void deleteAllEmployees() {
        employeeRepository.deleteAll();
    }

    public Employee updateEmployee(Employee oldEmp) {
        return employeeRepository.save(oldEmp);
    }
}
